# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ficha Digital API",
        "description": "API para el sistema de gestión de solicitudes y fichas digitales",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/promotores": {
            "get": {
                "tags": [
                    "Promotores"
                ],
                "summary": "Listar promotores activos",
                "description": "Lista todos los promotores que existen (existe = 1)",
                "operationId": "71c6ab80aeed859e054196eef6e3389e",
                "responses": {
                    "200": {
                        "description": "Lista de promotores activos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Promotores activos"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/promotores/buscar-legajo/{legajo}": {
            "get": {
                "tags": [
                    "Promotores"
                ],
                "summary": "Buscar promotor por legajo (IdProductorCp)",
                "description": "Busca un promotor por su legajo (IdProductorCp) que existe (existe = 1)",
                "operationId": "640826235426c21bcc0dab269762b96f",
                "parameters": [
                    {
                        "name": "legajo",
                        "in": "path",
                        "description": "Legajo del promotor (IdProductorCp)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "12345"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promotor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Promotor encontrado"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Promotor no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Promotor no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/promotores/buscar": {
            "get": {
                "tags": [
                    "Promotores"
                ],
                "summary": "Buscar promotores por apellido y/o nombre",
                "description": "Busca promotores por apellido y/o nombre que existen (existe = 1). Mínimo 3 caracteres para la búsqueda.",
                "operationId": "db1456eae063b7b6888057d1f8214c6d",
                "parameters": [
                    {
                        "name": "apellido",
                        "in": "query",
                        "description": "Apellido del promotor (mínimo 3 caracteres)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "García"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre del promotor (mínimo 3 caracteres)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Juan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promotores encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Promotores encontrados"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Debe proporcionar al menos apellido o nombre con mínimo 3 caracteres"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}